Add support for repo.yml; select type
This is a breaking change. There's a new repo-yml plugin that will
detect the presene of a repo.yml file and read the type out of it. If
that's present, the type for the project will be set to the type
specified in repo.yml.

Plugins now support a priority integer, lower being higher priority.
Assuming that there is no repo.yml, the type will be selected in order
of priority from the available plugins. The first plugin in priority
order which has returned true will be the type for the repo.

Change-type: major
Signed-off-by: Jack Brown <jack@resin.io>